Add a reclassification of known deterministic
This will allow for the cache to be used where possible. We should be
careful to only include strategies in this list for which we are sure.

I will add tests for this but waiting for #12 to go in so as to avoid
merge conflicts.

I'm thinking that for all the known strategies I can add a test like the
current `test_tit_for_tat` which checks classification and behaviour (as
described in #11).